# Conflicts:
#	dtm-query-execution-core/config/application-default.yml
#	dtm-query-execution-core/config/application-dev.yml
#	dtm-query-execution-core/config/application-local-debug.yml
#	dtm-query-execution-core/config/application-local.yml
#	dtm-query-execution-core/src/main/java/io/arenadata/dtm/query/execution/core/service/edml/impl/DownloadKafkaExecutor.java

+++ dtm-jdbc-driver/src/main/java/io/arenadata/dtm/jdbc/ext/DtmResultSet.java
@@ -206,7 +206,7 @@ public class DtmResultSet implements ResultSet {
     @Override
     public float getFloat(int columnIndex) throws SQLException {
         final Object value = this.getValue(columnIndex);
-        return value == null ? 0 : (Float) value;
+        return value == null ? 0 : ((Double) value).floatValue();
     }
 
     @Override
